+++ lib/bible_chichewa.dart
@@ -1,2 +1,8 @@
+
+/*
+ author: M2K Developments
+ github: https://github.com/m2kdevelopments
+ */
+
 library bible_chichewa;
 export 'src/main.dart';