+++ lib/src/main.dart
@@ -1,3 +1,8 @@
+/**
+ author: M2K Developments
+ github: https://github.com/m2kdevelopments
+ */
+
 import 'dart:convert';
 import 'package:flutter/services.dart';