+++ CHANGELOG.md
@@ -1,7 +1,24 @@
 ## [1.0.0] - 2023-08-23
 ### Added
-- New feature 1
-- New feature 2
+- 5 Functions to get Bible Content
+
+
+## [1.0.1] - 2023-08-23
 
 ### Fixed
 - Loading Bible
+
+
+## [2.0.0] - 2023-08-23
+
+### Restore Missing Bible Verses
+- Restore Bible Verses
+- Matthew 17:21.
+- Matthew 18:11.
+- Matthew 23:14.
+- Mark 7:16.
+- Mark 9:44 and 9:46.
+- Mark 11:26.
+- Mark 15:28.
+- Luke 17:36.
+- and more